tests/wpt/web-platform-tests/bluetooth/server/getPrimaryServices/gen-garbage-collection-ran-during-error-with-uuid.https.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Garbage Collection ran during a getPrimaryServices ' +
    'call that failed. Should not crash.'
const expected = new DOMException(
    'GATT Server is disconnected. Cannot retrieve services. (Re)connect first ' +
    'with `device.gatt.connect`.',
    'NetworkError');
let promise;

bluetooth_test(() => getEmptyHealthThermometerDevice()
    .then(({device}) => {
      promise = assert_promise_rejects_with_message(
          device.gatt.getPrimaryServices('health_thermometer'),
          expected);
      // Disconnect called to clear attributeInstanceMap and allow the
      // object to get garbage collected.
      device.gatt.disconnect();
      return runGarbageCollection();
    })
    .then(() => promise),
    test_desc);

</script>
